+++ handlers/users/registration.py
@@ -24,17 +24,24 @@ async def bot_start(message: types.Message):
         await message.answer(text=settings_worker.get_choose_languages_text(), reply_markup=await get_languages_keyboard("reg", ref_id))
         await message.delete()
     else:
-        text = languages_worker.get_text_on_user_language(message.from_user.id, "mainMenu")
-        last_menu = users_worker.get_last_menu(message.from_user.id)
-        if last_menu:
-            try:
-                await bot.delete_message(message.from_user.id, last_menu)
-            except:
-                pass
+        is_paying = users_worker.get_is_paying(message.from_user.id)
 
-        main_msg = await message.answer(text=text["mainMenu"], reply_markup=await get_main_keyboard(message.from_user.id))
-        users_worker.update_last_menu(message.from_user.id, main_msg.message_id)
-        await message.delete()
+        if not is_paying:
+
+            text = languages_worker.get_text_on_user_language(message.from_user.id, "mainMenu")
+            last_menu = users_worker.get_last_menu(message.from_user.id)
+            if last_menu:
+                try:
+                    await bot.delete_message(message.from_user.id, last_menu)
+                except:
+                    pass
+
+            main_msg = await message.answer(text=text["mainMenu"], reply_markup=await get_main_keyboard(message.from_user.id))
+            users_worker.update_last_menu(message.from_user.id, main_msg.message_id)
+            await message.delete()
+        else:
+            text = languages_worker.get_text_on_user_language(message.from_user.id, "blockMenuInfo")
+            await message.answer(text["blockMenuInfo"], show_alert=True)
 
 
 @dp.callback_query_handler(language_callback.filter(action="reg"))
